backend/requirements.txt
asyncio
asyncpg
datetime
fastapi
firebase
firebase_admin
pydantic
sqlalchemy
uvicorn
pySerial
numpy
scikit-learn
sympy